Update sample reads listing and items
* Pull data from `detail.reads` instead of `detail.files`
* Make rendering of `from` field optional. This is changing for SQL-based reads
* Change box group title to Reads
* Use thick font weight in items
