+++ client/src/js/samples/components/Files/Reads.js
@@ -3,9 +3,9 @@ import React from "react";
 import { connect } from "react-redux";
 import styled from "styled-components";
 import { BoxGroup, BoxGroupHeader } from "../../../base";
-import { SampleRawItem } from "./RawItem";
+import { ReadItem } from "./ReadItem";
 
-const SampleFilesRawTitle = styled.h2`
+const SampleReadsTitle = styled.h2`
     display: flex;
     justify-content: space-between;
 
@@ -14,15 +14,15 @@ const SampleFilesRawTitle = styled.h2`
     }
 `;
 
-export const SampleFilesRaw = ({ files, prefix }) => {
-    const fileComponents = map(files, (file, index) => (
-        <SampleRawItem key={file.name} {...file} prefix={prefix} suffix={index + 1} />
+export const SampleReads = ({ reads, prefix }) => {
+    const fileComponents = map(reads, (file, index) => (
+        <ReadItem key={file.name} {...file} prefix={prefix} suffix={index + 1} />
     ));
 
     return (
         <BoxGroup>
             <BoxGroupHeader>
-                <SampleFilesRawTitle>Raw Data</SampleFilesRawTitle>
+                <SampleReadsTitle>Reads</SampleReadsTitle>
                 <p>The input sequencing data used to create this sample.</p>
             </BoxGroupHeader>
             {fileComponents}
@@ -31,12 +31,12 @@ export const SampleFilesRaw = ({ files, prefix }) => {
 };
 
 export const mapStateToProps = state => {
-    const { id, files, name } = state.samples.detail;
+    const { id, reads, name } = state.samples.detail;
     return {
-        files,
+        reads,
         id,
         prefix: snakeCase(name)
     };
 };
 
-export default connect(mapStateToProps)(SampleFilesRaw);
+export default connect(mapStateToProps)(SampleReads);
